[RFC] "Reparameterize" random effects internally (#629)
* Simplify computation in the marginal likelihood by using a scaled
version of the random effects for internally. Hence, when a
NamedTuple of random effects is converted to Cartesian space, the
Cholesky factor of the covariance matrix is solver for such that
the Vector representation correspons to a standard normal distribution.
This greatly simplifies all likelihood expression as the variance of
the random effects no longer enter the expressions. It also allows
that the covariance of the random effects is singular without
breakdown of the likelihood calculation. Hence, this commit should
make estimation more robust.

Fixes #570
Fixes #599

* Define transform for Normal to allow spcifying multiple random
effect variables.

* Fix Poisson tests

* Fix simple_model.jl tests

* Fix simple_model_tdist.jl tests

* Fix diagnostics tests

* Fix information.jl tests

* Mark a test as broken as it seems like a degenerate problem from
inspecting the optimization trace

* Fix ηshrinkage. It now returns a named tuple.

* Fitxups after rebase

* Fix computation of the covariance of EBEs